Fix fanfics spider selector workings
Add quick fixes to selectors for fanfics crawling in order to get rid of
'list out of range' errors. Description usually is biggest by amount of
chars, so it's selected this way (no idea how to select it better as it
is in parent div and is just text so no specific element can be
selected with ::text apart from this parent div).